requirements.txt
altgraph==0.17.4
anyio==3.6.2
APScheduler==3.6.3
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
decorator==5.1.1
Django==4.2
django-rest-auth==0.9.5
djangorestframework==3.14.0
docopt==0.6.2
future==0.18.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
iniconfig==2.0.0
macholib==1.16.3
packaging==23.2
pipreqs==0.4.13
pluggy==1.3.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pytest==7.4.3
python-dotenv==1.0.0
python-telegram-bot==12.0.0
pytz==2023.3
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
Telethon==1.28.2
tornado==6.2
tzlocal==5.2
urllib3==1.26.15
yarg==0.1.9
